@@ -3,41 +3,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "connected-react-router": "^5.0.1",
-    "immutability-helper": "^2.9.0",
-    "node-sass": "^4.12.0",
-    "query-string": "^6.4.0",
+    "node-sass": "^4.13.0",
     "rc-slider": "^8.7.1",
-    "react": "^16.6.3",
-    "react-addons-css-transition-group": "^15.6.2",
-    "react-addons-update": "^15.6.2",
-    "react-collapse": "^4.0.3",
-    "react-cursor-position": "^3.0.3",
-    "react-dd-menu": "^2.0.2",
-    "react-debounce-input": "^3.2.0",
-    "react-dom": "^16.6.3",
-    "react-file-viewer": "^1.1.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-horizontal-timeline": "^1.5.3",
     "react-infinite-scroller": "1.2.4",
-    "react-motion": "^0.5.2",
-    "react-redux": "^5.1.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "^2.1.3",
-    "react-selectable-fast": "^2.2.0",
-    "react-sidebar": "^3.0.2",
-    "react-table": "^6.10.3",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "^3.2.0",
+    "react-table": "^7.0.0-beta.15",
     "react-toggle-switch": "^3.0.4",
-    "react-transition-group": "^4.0.0",
-    "redux": "^4.0.1"
+    "react-virtualized-auto-sizer": "^1.0.2",
+    "react-window": "^1.8.5"
   },
   "scripts": {
     "start": "set PORT=81 && react-scripts --max_old_space_size=4096 start",
     "analyze": "source-map-explorer 'build/static/js/*.js'",
     "build": "react-scripts --max_old_space_size=4096 build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "postinstall": "cp node_modules/pdfjs-dist/build/pdf.worker.min.js public/pdf.worker.min.js"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
