@@ -1,57 +1,35 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './Header.scss';
-import { connect } from 'react-redux';
-import history from '../../history';
+import { Link } from 'react-router-dom';
 import DigitalNorthSeaLogo from '../../assets/DigitalNorthSeaLogo';
 import Slider from 'rc-slider';
-import setClock from '../../actions/mapActions'
 import 'rc-slider/assets/index.css';
-export const initialTabId = { id: 0 };
-class Header extends Component {
+import { useStateValue } from '../../utils/state';
 
-  constructor(props){
-    super(props);
-    this.changeTime = this.changeTime.bind(this);
-  }
+const marks = {
+  1975: 1975,
+  1985: 1985,
+  1995: 1995,
+  2005: 2005,
+  2015: 2015,
+  2025: 2025,
+  2035: 2035
+};
 
-  changeTime(value){
-    this.props.changeClock(value);
-  }
+const Header = () => {
+  const [, dispatch] = useStateValue();
+  return (
+    <div className="header-container">
+      <Link className="logo-container" to={""} >
 
-  render() {
-
-    const marks = {
-      1975:1975,
-      1985:1985,
-      1995:1995,
-      2005:2005,
-      2015:2015,
-      2025:2025,
-      2035:2035            
-    };
-
-    return (
-      <div className="header-container">
-        <div className="logo-container" onClick={() => {
-          history.push("")
-        }}>
-          <DigitalNorthSeaLogo></DigitalNorthSeaLogo>
-        </div>
-        <Slider min={1975} max={2035} marks={marks} defaultValue={new Date().getFullYear()} onChange={this.changeTime}></Slider>
-        <div className="spacer">
-        </div>
+        <DigitalNorthSeaLogo></DigitalNorthSeaLogo>
+      </Link>
+      <Slider min={1975} max={2035} marks={marks} defaultValue={new Date().getFullYear()} onChange={e => dispatch({ type: "changeYear", year: e })}></Slider>
+      <div className="spacer">
       </div>
+    </div>
 
-    );
-  }
+  );
 }
 
-function mapDispatchToProps(dispatch) {
-  return {
-      changeClock: (year) => {
-          dispatch(setClock(year))
-      }
-  }
-}
-
-export default connect(null, mapDispatchToProps)(Header)
+export default Header;
